requirements.txt
gunicorn==20.1.0
numpy>=1.23.5
flask>=2.2.2
tensorflow>=2.3.0
tensorflow-gpu>=2.3.0
pillow>=9.3.0